Moved authentication backend to seperate directory
