+++ backend/authentication/controllers/userController.go
@@ -6,9 +6,9 @@ import (
 	"log"
 	"net/http"
 	"strconv"
-	"tictactoe/database"
-	helper "tictactoe/helpers"
-	"tictactoe/models"
+	"tictactoe/authentication/database"
+	"tictactoe/authentication/helpers"
+	"tictactoe/authentication/models"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -82,7 +82,7 @@ func Signup() gin.HandlerFunc {
 		user.Updated_at, _ = time.Parse(time.RFC3339, time.Now().Format(time.RFC3339))
 		user.ID = primitive.NewObjectID()
 		user.User_id = user.ID.Hex()
-		token, refreshToken, _ := helper.GenerateAllTokens(*user.Username, *user.First_name, *user.Last_name, *&user.User_id)
+		token, refreshToken, _ := helpers.GenerateAllTokens(*user.Username, *user.First_name, *user.Last_name, *&user.User_id)
 		user.Token = &token
 		user.Refresh_token = &refreshToken
 
@@ -133,13 +133,13 @@ func Login() gin.HandlerFunc {
 			return
 		}
 
-		token, refreshToken, _ := helper.GenerateAllTokens(
+		token, refreshToken, _ := helpers.GenerateAllTokens(
 			*foundUser.Username,
 			*foundUser.First_name,
 			*foundUser.Last_name,
 			*&foundUser.User_id,
 		)
-		helper.UpdateAllTokens(token, refreshToken, foundUser.User_id)
+		helpers.UpdateAllTokens(token, refreshToken, foundUser.User_id)
 		err = userCollection.FindOne(ctx, bson.M{"user_id": foundUser.User_id}).Decode(&foundUser)
 		if err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
@@ -219,7 +219,7 @@ func GetUser() gin.HandlerFunc {
 	return func(c *gin.Context) {
 		userId := c.Param("user_id")
 
-		if err := helper.MatchUserTypeToUid(c, userId); err != nil {
+		if err := helpers.MatchUserTypeToUid(c, userId); err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 			return
 		}
