+++ backend/authentication/helpers/authHelper.go
