+++ backend/websocket/player.go
@@ -1,0 +1,1 @@
+package websocket
